- 20-08-2004 changed: mplayers SeekTime() now is in ms for better seek accuracy for .cue files etc.   - 20-08-2004 changed: Updated the mplayer interface to handle the new DTS decoding capabilities.   - 20-08-2004 changed: Removed AC3/DTS pass through options.  Will always passthrough unless it can't (non-48kHz),                         or the user chooses to output stereo files to all speakers.   - 20-08-2004 changed: Audio options now set the MS Dash settings for AC3/DTS.   - 20-08-2004 changed: Music Database Reorg now Clean Music Database - just removes moved/deleted songs.   - 20-08-2004 changed: Database updates for My music to better streamline queries + size.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@1956 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/ModPlayer.h
@@ -14,7 +14,7 @@ public:
 	virtual void		RegisterAudioCallback(IAudioCallback* pCallback);
 	virtual void		UnRegisterAudioCallback();
 
-	virtual bool		openfile(const CStdString& strFile, int iStartTime);
+	virtual bool		openfile(const CStdString& strFile, __int64 iStartTime);
 	virtual bool		closefile();
 	virtual bool		IsPlaying() const;
 	virtual void		Pause();
